requirements.txt
# RPA Framework Dependencies
pyautogui
opencv-python
pywin32
pillow
numpy
psutil
keyboard
mouse
pyyaml
colorlog
pytest
pytest-cov
easyocr
requests